requirements.txt
marker-pdf
boto3 
torch
requests
python-dotenv
